+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Toybox Changelog
+## v2.1.0 (22/05/2023)
+* Modularize components.
+
+## v2.0.0 (15/05/2023)
+* Split theme into two packages: toybox/theme and toybox/core.
+* New enhancements to tweak more of WordPress.
+
 ## v1.1.1 (19/03/2023)
 * Adds this changelog.
 * Updates npm dependencies (flexya, @glidejs/glide, sass).